Move ExperimentDetails and PlateSizeForm to activity folder
