+++ ui/src/components/activity/ActivityDetails/Version.js
